Updated variable types and refactoring

+++ src/main/kotlin/Util/Time.kt
@@ -16,8 +16,8 @@ class Time
     var timer = System.currentTimeMillis()
 
 
-    fun currentTime(): Long {
-        return System.nanoTime();
+    fun currentTime(): Double {
+        return System.nanoTime().toDouble();
     }
 
 }
